Implemented directional GeoCords, added barcode API, prepared Security, and fixed warnings/memory management

+++ src/main/java/com/congapp/ShelterInv/service/DonatorService.java
@@ -6,26 +6,27 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
-import org.springframework.data.geo.Circle;
-import org.springframework.data.geo.Distance;
-import org.springframework.data.geo.Metrics;
-import org.springframework.data.geo.Point;
+// import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+// import org.springframework.security.crypto.password.PasswordEncoder;
 
 import com.congapp.ShelterInv.dao.DonatorDao;
 import com.congapp.ShelterInv.model.Donator;
-import com.congapp.ShelterInv.model.Shelter;
 
 @Service
 public class DonatorService {
 
-    private final DonatorDao donatorDao;
+    private final DonatorDao donatorDao; 
+    // private final PasswordEncoder passwordEncoder;
 
     @Autowired
     public DonatorService (@Qualifier("donDao") DonatorDao donatorDao){
         this.donatorDao = donatorDao;
+        // passwordEncoder = new BCryptPasswordEncoder();
     }
     
     public Donator addPerson (Donator donator){
+        // String password = passwordEncoder.encode(donator.getPassword());
+        // donator.setPassword(password);
         return donatorDao.insert(donator);
     }
 
@@ -33,32 +34,30 @@ public class DonatorService {
         return donatorDao.findAll();
     }
 
-    public List<Shelter> findByDistance(float latitude, double longitude, int distance){
-        Point baseCords = new Point (latitude, longitude);
-
-        Distance radius = new Distance (distance, Metrics.MILES);
-
-        Circle area = new Circle(baseCords, radius);
-
-        Query query = new Query();
-        query.addCriteria(Criteria.where("address.geoLocation").withinSphere(area));
-
-        return MongoOperations.find(query, Shelter.class);
-
     public Optional<Donator> getPersonById(String id){
-        return donatorDao.findById(id);
+        if (donatorDao.findById(id).isPresent()){
+            return donatorDao.findById(id);
+        }
+        return null;
     }
 
     public void removePerson (String id){
-        donatorDao.deleteById(id);
+        if (donatorDao.findById(id).isPresent()){
+            donatorDao.deleteById(id);
+        }
     }
 
     public void changePersonName (String id, String name){
-        donatorDao.findById(id).get().setName(name);
+        if (donatorDao.findById(id).isPresent()){
+            donatorDao.findById(id).get().setName(name);
+        }
     }
 
     public String getPassword (String id){
-        return donatorDao.findById(id).get().getPassword();
+        if (donatorDao.findById(id).isPresent()){
+            return donatorDao.findById(id).get().getPassword();
+        }
+        return null;
     }
 
 }
